+++ tests/structs.rs
@@ -82,8 +82,5 @@ fn recursive_struct() {
 fn empty_struct() {
     #[derive(Miniconf, Default)]
     struct Settings {}
-    assert!(Settings::iter_paths::<1, String>("/")
-        .unwrap()
-        .next()
-        .is_none());
+    assert!(Settings::iter_paths::<String>("/").next().is_none());
 }